Improve manual_control with synced multicam recorder
I struggled with rendering a multicamera setting with high traffic (200 cars 200 pedestrians) on our good GPU-CPU server as it was lagging and consequently the physics calculations and autopilot calculations also deteriorated which led to a terrible experience. I opened an issue about this (https://github.com/carla-simulator/carla/issues/2617) and I found that many others had been struggling with this. 
Then I realized the solution is in setting the simulation time to fixed time setting. I was confused regarding the use of fixed _delta_time and synchronous mode until I realized the order of the game loop, which is as follows: world.tick() sending the tick to the server and blocking!, then -> on_tick() callbacks get called and sensor.listen() callbacks get called but these two's order might change, then returning from world.tick()! During these you can run the client loop on any fps if you separate the world.tick()-ing to another thread which is what I did. The important thing is that when you want to save multiple images from multiple sensors, it is better to wait for to write them to disk and then continue with the next server tick otherwise you will lose images or get bad images. For this multi-thread synchronization I used semaphore which I think is the mose appropriate solution. I have seen using queues in client_bounding_boxes.py, but I think that adds unnecessary complexity.
I think it is also important to explain this solution to newcomers because as you saw others struggled with slow fps, but if the rendering is controlled well like in this solution you can get beautiful results. I used 10 cameras for 5 stereo sides and I got beautiful results with epic rendering! :)
I also found out that the CarlaUE4.sh --fps and --benchmark settings didn't do anything, and I think they should be removed from documentation, because this can be controlled from the python api, but I won't open another issue for this.